+++ xalpha/__init__.py
@@ -35,12 +35,13 @@ from xalpha.universal import (
 from xalpha.provider import show_providers, set_proxy
 from xalpha.toolbox import (
     PEBHistory,
-    Compare,
     IndexPEBHistory,
     FundPEBHistory,
     SWPEBHistory,
     StockPEBHistory,
     TEBHistory,
+    Compare,
+    OverPriced,
     QDIIPredict,
     RTPredict,
     set_holdings,